pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sparrow-agent"
version = "1.0.0"
description = "Sparrow Enterprise Server Installation Agent (CLI)"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "requests>=2.31.0",
]

[project.scripts]
sparrow-agent = "main:main"

[tool.setuptools.packages.find]
include = ["agent*"]
